Refactored with auth.py
